rename : locker 및 jimtype 관련 DTO를 각각의 패키지로 분리
