+++ src/main/java/com/airbng/controller/LockerController.java
@@ -7,6 +7,8 @@ import com.airbng.dto.locker.LockerTop5Response;
 import com.airbng.service.LockerService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
