+++ src/main/java/com/airbng/dto/locker/LockerSearchResponse.java
@@ -1,4 +1,4 @@
-package com.airbng.dto;
+package com.airbng.dto.locker;
 
 import lombok.*;
