+++ src/main/resources/mappers/LockerMapper.xml
@@ -49,7 +49,7 @@
         <result column="url"            property="url" />
 
         <!-- 중첩 짐타입 매핑 -->
-        <collection property="jimTypeResults" ofType="com.airbng.dto.JimTypeResult">
+        <collection property="jimTypeResults" ofType="com.airbng.dto.jimType.JimTypeResult">
             <id     column="jimtype_id"    property="jimTypeId" />
             <result column="type_name"     property="typeName" />
         </collection>
@@ -98,10 +98,13 @@
             useGeneratedKeys="true" keyProperty="lockerId" keyColumn="locker_id">
         INSERT INTO Locker (
             is_available, locker_name, member_id, address,
-            address_english, address_detail, latitude, longitude
+            is_available, locker_name, member_id,
+            address, address_english, address_detail,
+            latitude, longitude
         ) VALUES (
-            #{isAvailable}, #{lockerName}, #{keeper.memberId}, #{address},
-            #{addressEnglish}, #{addressDetail}, #{latitude}, #{longitude}
+            #{isAvailable}, #{lockerName}, #{keeper.memberId},
+            #{address}, #{addressEnglish}, #{addressDetail},
+            #{latitude}, #{longitude}
         )
     </insert>
 
@@ -127,9 +130,7 @@
 
     <!-- 4. 보관소-짐타입 연결 -->
     <insert id="insertLockerJimTypes">
-        INSERT INTO LockerJimType (
-            locker_id, jimtype_id
-        )
+        INSERT INTO LockerJimType (locker_id, jimtype_id)
         <foreach collection="jimTypeIds" item="jimtypeId" separator="," open="VALUES" close="">
             (#{lockerId}, #{jimtypeId})
         </foreach>
@@ -158,15 +159,15 @@
         )
     </select>
 
-  
+
     <!-- ResultMap for Locker + JimType -->
-    <resultMap id="LockerResultMap" type="com.airbng.dto.LockerPreviewResult">
+    <resultMap id="LockerResultMap" type="com.airbng.dto.locker.LockerPreviewResult">
         <id column="locker_id" property="lockerId"/>
         <result column="locker_name" property="lockerName"/>
         <result column="is_available" property="isAvailable"/>
         <result column="url" property="url"/>
         <result column="address" property="address"/>
-        <collection property="jimTypeResults" ofType="com.airbng.dto.JimTypeResult">
+        <collection property="jimTypeResults" ofType="com.airbng.dto.jimType.JimTypeResult">
             <id column="jimtype_id" property="jimTypeId"/>
             <result column="type_name" property="typeName"/>
         </collection>
